[tests][mmp] Convert old mmp calendar regression tests to unit tests
