+++ tests/linker/mac/link all/LinkAllTest.cs
@@ -9,9 +9,21 @@ namespace LinkAllTests
 	[Preserve (AllMembers = true)]
 	public class LinkAllTest
 	{
-		[Test]
-		public void Existence ()
+
+		static void Check (string calendarName, bool present)
 		{
+			var type = Type.GetType ("System.Globalization." + calendarName);
+			bool success = present == (type != null);
+			Assert.AreEqual (present, type != null, calendarName);
+		}
+
+		[Test]
+		public void Calendars ()
+		{
+			Check ("GregorianCalendar", true);
+			Check ("UmAlQuraCalendar", false);
+			Check ("HijriCalendar", false);
+			Check ("ThaiBuddhistCalendar", false);
 		}
 	}
 }
