+++ tests/linker/mac/link sdk/LinkSdkTest.cs
@@ -9,9 +9,21 @@ namespace LinkSdkTests
 	[Preserve (AllMembers = true)]
 	public class LinkSdkTest
 	{
-		[Test]
-		public void Existence ()
+		static void Check (string calendarName, bool present)
 		{
+			var type = Type.GetType ("System.Globalization." + calendarName);
+			bool success = present == (type != null);
+			Assert.AreEqual (present, type != null, calendarName);
+		}
+
+		[Test]
+		public void Calendars ()
+		{
+			Check ("GregorianCalendar", true);
+			// because project options enabled them
+			Check ("UmAlQuraCalendar", true);
+			Check ("HijriCalendar", true);
+			Check ("ThaiBuddhistCalendar", true);
 		}
 	}
 }
