+++ tests/linker/mac/link sdk/link sdk-mac.csproj
@@ -31,6 +31,7 @@
     <LinkMode>SdkOnly</LinkMode>
     <XamMacArch></XamMacArch>
     <AOTMode>None</AOTMode>
+    <I18n>mideast,other</I18n>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|x86' ">
     <DebugType></DebugType>
