--- tests/mmptest/regression/link-preserve-calendar-2/Info.plist
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>CFBundleDisplayName</key>
-	<string>link-preserve-calendar-2</string>
-	<key>CFBundleIdentifier</key>
-	<string>com.your-company.linkpreservecalendar2</string>
-	<key>CFBundleName</key>
-	<string>link-preserve-calendar-2</string>
-	<key>CFBundleVersion</key>
-	<string>1</string>
-	<key>LSMinimumSystemVersion</key>
-	<string>10.7</string>
-	<key>NSMainNibFile</key>
-	<string>MainMenu</string>
-	<key>NSPrincipalClass</key>
-	<string>NSApplication</string>
-	<key>LSUIElement</key>
-	<string>1</string>
-</dict>
-</plist>
